+++ DisguiseUnityRenderStream/Editor/DisguiseRenderStreamBuild.cs
@@ -1,4 +1,3 @@
-#if UNITY_EDITOR
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
@@ -6,8 +5,9 @@ using UnityEngine.SceneManagement;
 
 namespace Disguise.RenderStream
 {
-    partial class DisguiseRenderStream  : UnityEditor.Build.IPreprocessBuildWithReport
+    class DisguiseRenderStreamBuildProcessor  : UnityEditor.Build.IPreprocessBuildWithReport
     {
+        static ManagedSchema schema;
         public int callbackOrder
         {
             get { return 0; }
@@ -116,4 +116,3 @@ namespace Disguise.RenderStream
         }
     }
 }
-#endif
