+++ DisguiseUnityRenderStream/Runtime/DisguisePluginDataTypes.cs
