+++ DisguiseUnityRenderStream/Runtime/PluginEntry.cs
